pdc: don't check for hard reset with non PD partners
When a non PD partner is connected, there's no need to check for hard
resets as neither the PDC nor the partner should generate a hard reset
on the CC lines.

This fixes an issue where a false hard reset was recorded and caused the
PDC subsystem to get stuck in the SNK_TYPEC_ATTACHED_DEBOUNCE state.

BUG=b:339511158
TEST=Perform PDC firmware update. Verify that depthcharge is able to
read the PDC firmware version after the update.

Change-Id: Ifbcf2a3eeeb0f276b73d33ca9d234110cea12dd5