zephyr: watchdog: npcx: change the way to set wdt warning time
In the PR:72506, the pre-scalar of the watchdog timer can be various
in order to support a wider range of the watchdog timeout.
We cannot use the WDT_NPCX_DELAY_CYCLES to set the watchdog leading
warning time because the time of one watchdog count is not necessary
32ms after that change. This CL uses a new configuration
CONFIG_WDT_NPCX_WARNING_LEADING_TIME_MS whose unit is ms to set
the watchdog leading warning time instead.
This CL has the dependency on PR:
https://github.com/zephyrproject-rtos/zephyr/pull/72506

BRANCH=none
BUG=b:309557100
TEST=
With default Watchdog setting, use console command:
  "waitms 1600" - leading warning message printed -> EC resets.
  "waitms 1100" - leading warning message printed.
With "CONFIG_PLATFORM_EC_WATCHDOG_PERIOD_MS = 10000":
  "waitms 10000" - leading warning message printed -> EC resets.
  "waitms 9500" - leading warning message printed.

Change-Id: I04e6e11defe2593a176da94d13e31b3e78bc8226
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>