board/hyperdebug: Optimizing bitbanging
Interrupt handler composed just-in-time from snippets of hand written
assembly.  This makes the critical path faster, as it does not need to
iterate.

It also allows functional improvements, by setting multiple pins on the
same IO bank simultaneously.  (Detecting and handling such cases at
runtime would be prohibitively slow, but when the determination happens
at setup time, and the result is a piece of linear machine code, it
makes sense.)

BUG=b:341302705
TEST=TBD

Change-Id: I3659c0bdcc5971e26d323206f6583ca651f79ec4