rollback: Fix stack leaking when no rollback region is initialized
When no rollback region exists on flash, get_latest_rollback() will not
fully initialize it's argument (only 'secret' is cleared). On the other
side, rollback_update() assumes that the structure was initialized.
As a result, new rollback region can have invalid data.

Please note that this only affects tests. Firmware is not affected,
because the binary contains rollback initial data.

BUG=b:340215803
TEST=twister -v -i -T zephyr/test/fingerprint --toolchain llvm
TEST=twister -v -i -T zephyr/test/fingerprint --toolchain host
TEST=twister -v -i -T zephyr/test/rollback --toolchain llvm
TEST=twister -v -i -T zephyr/test/rollback --toolchain host

Change-Id: Ib69e0d55c942e08268e1daf71916a7fae908cdd8