pdc: Fix SET_UOR/SET_PDR command usage in snk/src states
This CL fixes the both the SET_UOR and SET_PDR command usage within
the PDC subsys API. For both commands bits 1:0 are used to request a
data/power role and both of these bits can't be clear, or the PDC will
treat it as an illegal command and return an ping error status.

BUG=b:326030089
TEST=Verifed on Brox that SET_PDR that's sent in both SNK_ATTACHED and
SRC_ATTACHED states is successful. Verified that SET_UOR works in
cases where an alternate mode is not already been entered.

Change-Id: Ic7442f825d489999d009f6e1b9f4a099d9bffc0c