pdc: Modify effect of SET_ACTIVE_CHARGE_PORT flag
The flag SNK_POLICY_SET_ACTIVE_CHARGE_PORT is set via a call from the
charge manager whenever there is some change to the active charge
port. Prior to this CL, this flag being set was setting the
SNK_ATTACHED local state to SNK_ATTACHED_GET_PDOS. However, the charge
manager will already have all of the updated PD contract info from all
ports prior to it calling `set_active_charge_port()`. So it isn't
necessary to back all the way up to SNK_ATTACHED_GET_PDOS when
SET_ACTIVE_CHARGE_PORT is set. Rather, it's only necessary to set the
local state to SNK_ATTACHED_SET_SINK_PATH.

This will reduce the number of PDC commands that are being sent to the
PDC when a port partner is attached/detached and both ports are in the
SNK_ATTACHED state.

BUG=b:340577355
TEST=Verifed that charging still works for both ports as the port
partners are attached/detached.

Change-Id: I658da7888575a93a630e0b8dc3de1bef07ee0c28