cbi: always check cros_cbi_get_fw_config return value
Fix all instances that were ignoring the return value of
cros_cbi_get_fw_config. The function now returns an error if the
underlying data has not been initialized, this changes all occurences of
this call to detect the condition and fail whatever they are doing
rather than continuing silently with whatever value matched 0.

For the Vivaldi keyboard specifically this means that the condition
would disable vivaldi entirely rather than defaulting to the first
config, making the condition detectable in qualification even if the
target device so happen to match the first config.

BUG=b:340187957
TEST=cq dry run

Change-Id: If375c24be23d668c69e0e7d30db5726e78a83383