chip/stm32: Configure DBGMCU for STM32L5
It seems that when I originally added support for the L5 series.  I
disabled some common STM32 code for setting up DBGMCU, rather than
figuring out how to make it work on the L5.

This CL adds a declaration of STM32_DBGMCU_BASE, in order to configure
the chip to stop all timers in cases when the core is stopped by a
debugger.

This is a very small feature, but I would rather get rid of special
cases or omissions for L5.

BUG=none
TEST=make BOARD=hyperdebug  (boots to serial console)

Change-Id: I3fccab0c578855d34e68ad7462968e71987a7d20