chip/stm32: Enable instruction cache for STM32L5
Similar to existing code in clock-f.c, which enables caches for the
STM32F4 family, this CL adds code to enable the instruction cache for
the STM32L5 family.

With this CL, the timer interrupt handle used for bitbanging on
HyperDebug completes in half the time it used to, allowing faster
waveforms to be synthesized.

BUG=none
TEST=make BOARD=hyperdebug

Change-Id: I2394f259e5d0ed6e6404f45f4b811546c49dd70f