cortex-m: Implement IRQ handler enter/exit in ASM
Reimplement exiting from IRQ handlers in assembly to have more control
over task switching process. This will eliminate problem with compiler
optimizations that may mess up with task switching. (E.g. current code
was working on EC only because there is optimization that if function
call is last instruction in C code, then it can be replaced with "b"
instruction and stack is unwinded before the jump)

Now svc_handler calls svc_scheduler_handler. The later is performing
actions of previous svc_handler, but returns 0 if task doesn't need to
be switched or previous value of current task if current_task was
changed. New svc_handler is implelemented in such way to prevent
clobbering any registers in r4-r11 range. It interprets value returned
by svc_scheduler_handler and calls __switchto if needed.

The task_resched_if_needed is also implemented in the way preventing
clobbering r4-r11 registers. It is caller responsibility to unwind stack
correctly. One instance of that is npcx and mchp watchdog which does it
correctly. Another instance is DECLARE_IRQ. It is rewritten in assembly
to control which registers are used.

BUG=b:342114880
TEST=Get any board with Cortex-M EC and see if task switching is working

Change-Id: Iec9c777acb7cd7a825deba83fb25157c419e7199