chip/stm32: Configurable UART clock source for STM32L5
First, this CL changes the default behavior on STM32L5 to match that of
STM32L4.  That is, the serial console UART will be based on the HSI16
oscillator, meaning that there will be no glitches on the UART console,
even when switching the source of the core clock.

Any other UARTS used for forwarding to USB endpoints will be based on
the peripheral clock, meaning that they can better approximate
non-standard baud rates, if so requested via USB_USART_SET_BAUD.

Additionally for STM32L5, new CONFIG_ defines allow a board to opt to
use the fixed HSI16 clock or SYSCLK as source for the USB UARTS, instead
of the default APB PCLK.  This allows each board to decide the tradeoff.

For HyperDebug, the plan is to use the fastest possible SYSCLK, in order
to be able to use 1.5Mbaud communication for fast bootstrapping of
OpenTitan.

BUG=b:342243140
TEST=make BOARD=hyperdebug

Change-Id: I9d938d6133eaa1ab0fa477e356e0a4c94547ff45