board/hyperdebug: Improve USB serial forwarding
One of the HyperDebug UARTs is commonly connected to the GSC console.
In order to speed up serial rescue, we plan on running this port at up
to 1.5Mbaud (unlike the three other forwarded UARTs, which will remain
at 115.2kbaud).

In order to support such fast speeds, this CL increases the size of the
USB buffer, in order to allow full 64-byte packages on the USB bus.
With the previous USB packet limit of 16 bytes, HyperDebug was not able
to saturate the UART at 1.5Mbaud.

Also, use 110MHs SYSCLK as base frequency for the UART IP blocks,
instead of a separate 16MHz HSI clock.  This allows high baud rates to
be realized with better accuracy.

BUG=b:342243140
TEST=make tast TAST_EXPR=gscdevboard.GSCUARTForward  (from ti50/common)

Change-Id: I0ff6cb9524ff7cea55277b78aee448bbf4324dd4