fpsensor: Simplify build.mk
By default, the fpsensor objects should only be included in RW. However,
the test build is different. The "emulator" runs tests against RO, so we
need to include the files in RO when building for it. The board tests
(on-device tests) build the tests for RO too (but we run them in RW).

BUG=b:336312700
TEST=CQ passes

Change-Id: Id9c469865a7fd77c6f76c7c28e163731ce66871a