ISL923x: Configure AC/DC Prochot current
If the PSYS comparator is enabled, AC/DC Prochot settings will be used
to keep the current below the setpoints. However they were being left at
the default values which are too low.

Default the input current to 6A, which gives room for 5A adapters, and
set the battery current to 12A. Individual projects may need to
overwrite these values but the default battery current of 4A is much too
low.

BUG=b:343370888
TEST=make buildall
TEST=zmake build brox
TEST=Probe prochot while running tast benchmarks, it should not assert

Change-Id: I62915126e1d85ca9ed6f26b6e8f8289a77162097