Wugtrio: Configure separate drv_data for accel and gyro on the LSM6DSM
The drv_data for the accelerometer and gyroscope on the LSM6DSM is configured
to be the same, so their calibration data will be identical after calibration.
This will cause the device to be unable to rotate. Therefore, configure
separate drv_data for the accelerometer and gyroscope on the LSM6DSM.

BUG=b:345144201
TEST=zmake build wugtrio and do calibrate operation,
     the DUT can rotate normally.

Change-Id: Ifc3e9831c667de2c1e95e6b5c293888d864c05d6