pdc: Cache VBUS voltage and identity VDOs
Currently, GET_VBUS_VOLTAGE and GET_VDO is rapidly sent to the PDC when
we have a Servo attached (and potentially in other scenarios as well).
This results in PDC traffic every 50-100ms which generally slows down
PDC functionality.

GET_VDO can return 0s for VDOs which matches the uninitialized value and
keeps erroring. Update the uninitialized value to -1u and reset this on
unattached.

GET_VBUS_VOLTAGE is queried frequently so cache the result and only
re-query every 500ms. The timeout for this gets reset on every unattach
and power level change notification from the PDC.

BUG=b:344266168
TEST=./twister -T zephyr/test/pdc and manual test on Brox

Change-Id: I72e6d683b753719f3edccedfc89082127ddb5195