zephyr: flash: npcx: set_status should return error if HWWP is enabled
When the WP_IF in the DEV_CTL4 register
(i.e. /WP pin of the internal flash) is enabled and the SRP0 bit
in the status register 1 of the internal flash is set, writing to the
status registers is invalid. The flash_set_status function currently
doesn't check it and return success in any way. This CL changes the
flash_set_status function to return an error code if the SPI flash
hardware write protect is enabled.

It is the equivalent change to the CL:4933562 required by the FPMCU
unit test.

BUG=b:309557100
BRANCH=none
TEST=make sure flash_set_status() return EC_ERROR_ACCESS_DENIED if
Hardware Protection function is enabled.
LOW_COVERAGE_REASON=this is the chip-specific hardware driver and
cannot be fully tested via the emulator.

Change-Id: I3f11c94fed7581e8c7a6d0746d90fe8972f017a5
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>