cmake: Depend on libclang_rt.builtins-armv7m.lib by full path
The directory in which libclang_rt.builtins-armv7m.lib resides is no
longer automatically added to the search path in newer Clang
versions. As a result, -lclang_rt.builtins-armv7m fails to find it. With
this change, we instead pass the full path to the library as an input
file. This will cause the library to be used for symbol resolution,
without adding its location to the search path.

BUG=b:346309204
TEST=build chromeos-base/chromeos-fpmcu-bloonchipper-unittests

Change-Id: I513e702275790799e9c561affecafb7969b669de