pdc: ti: Add live/cached PD chip info support
The TI driver was ignoring the live parameter to `tps_get_info`. This
means that the driver was initiating a live, asynchronous read operation
against the PDC, even if the caller requested cached chip data. This is
problematic because the driver was hanging on to a pointer into the
caller's memory. When the async PDC read finished, that pointer was no
longer valid as the caller had long exited. Writing out the received
chip info to that location actually clobbered the stack (most likely the
host command task's stack)

This CL implements full support for the `live` parameter in the chip info
API (we need this for Depthcharge FW updates later). Also prefetch the
chip info during driver init to seed the cached value.

BUG=b:346607670
BUG=b:346412014
TEST=pdc info {0|1} {0|1}; ectool pdchipinfo {0|1} {0|1}
LOW_COVERAGE_REASON=Coverage to be handled in b/345292002

Change-Id: If6906da83fecde21d122ff4dfc9d20d59d8fbeb0