zephyr: add an option to use CrosEC sbrk function
Add an option to use the CrosEC version of the sbrk function.
It is used by default when newlib is present.

The CrosEC version, as opposite to the one in libc-hooks for newlib,
doesn't give memory at the end of RAM, where the jump data is stored.

BUG=b:324574796
TEST=make buildall && zmake build -a
LOW_COVERAGE_REASON=Tests added in CL:5639700 as a part of b:324574796

Change-Id: I674275962b6da006876cb06ee3542ce452306fd8