zmake: add python to update the usb configuaration string descriptor
The EC firmware version is obtained from the USB configuration
filesystem during the automatic firmware update. Additionally, the USB
configuration string is defined as a KConfig option in the Zephyr
framework. This commit adds a python script and updates the definition
to the EC firmware version string during compile time.

BUG=b:327546783
BUG=b:347662534
TEST=zmake build -a

Change-Id: I6a2c8845dbd9fba67a37bcfd9005dd3e58aeb1f0