guybrush: Use default keyboard_scan_config
Remove keyboard_scan_config override for guybrush and use the common
default config. These overrides have been copy pasted from older
generations and not actually tuned for this board. In particular the
short scan_period_us results in the keyscan task to consuming a large
portion of the avaiable task time when key scanning is active.
This change masks the key at column 0, row 5. This is an
"unknown"/unused key.

BUG=b:346492855
TEST=Keyboard works normally on dewatt. Keyscan task consumes less time.

Change-Id: I5beedbe67da3495f55f436df5b90d26117942f2d