brox: switch to the upstream keyboard driver
Switch all Brox based projects to use the upstream keyboard driver
stack. This has been tested on a P1 and all the feature used on the
board (boot keys, runtime keys etc) seems to be working fine. Ran some
tast tests as well with the new stack on a Rex board (that's what I have
working here for tests) and they seem to work as well.

Firmware space requirements are same in RAM and equivalent or slightly
higher in Flash depending on what shell commands are active in the
config.

BUG=b:274443029
TEST=ran on a brox device
TEST=build only on lotso and greenbayupoc
TEST=cq dry run
TEST=ran few tast tests (on rex though)

Change-Id: Idff66d5bd0efbc8c531b66df4dc91b252515fed2