test: improve stdlib test
Improve the memmove test inside the stdlib test.

Verify correctness of the memmove function when the buffers overlap.

The previous version of TEST_ASSERT_ARRAY_EQ wasn't actually checking
the memmove. It was passing only because the big number of calls e.g.
buff = {1, 2, 3, 4, 5, 6}
memmove(buf + 2, buf, 4) should result in
buf = {1, 2, 1, 2, 3, 4}
so TEST_ASSERT_ARRAY_EQ(buf, buf + 2, 4) is not a good check.

BUG=b:324574796
TEST=make buildall

Change-Id: I35e1ac22b8d16a8f7bebbb26bb8bc6962bdaa167