zmake: Add support for optional modules
Some projects require private modules for full functionality, but
also can be built without them (public build). We can't add private
modules to 'modules' list, because 'prune_modules()' will raise an
exception if the module can't be found.

The problem is solved by adding 'optional_modules' list which is
the same as 'modules', but prints warning when optional module is
not available

BUG=b:314208406
TEST=zephyr/zmake/run_tests.sh

Change-Id: Ic68905bbeb05a3d2b4b7d742b89817f74786c47c