brox: Override function to disable keyboard, when in Tablet Mode
When Dut is in tablet mode and is suspended, onboard keyboard scan event
must be disabled. Brox sku is a convertible and hence, the function -
lid_angle_peripheral_enable() must be overridden to ensure that lid open
scenario is ignored, when dut is entered into tablet mode by command
method.

BUG=b:346270509
BRANCH=NONE
TEST=Enter Tablet mode by console command-tabletmode on & suspend dut.
On pressing onboard keyboard, dut doesn't wake & continues to be in suspend.

Change-Id: I646ebc4f072937d4971d015fdb71818ce68063e1
Signed-off-by: Poornima Tom <poornima.tom@intel.com>