lid_angle: Handle tabletmode scenario in common function
In the case of tabletmode, when system is suspended, on board keyboard
scanning is not required to be enabled. This CL helps in removing
duplicate code, where function-lid_angle_peripheral_enable() is made
more generic to more boards. The cl takes care of tabletmode entry for
convertibles, where lid angle can be ignored. The function -
lid_angle_peripheral_enable() is still continuing to be overridable
inorder to ensure any specific scenario that needs to be taken care in
some boards.

BUG=b:346270509
LOW_COVERAGE_REASON=Coverage to be handled in b/350787417
TEST=Build all.Tested on Brox

Enter Tablet mode by console command-tabletmode on & suspend dut. On
pressing onboard keyboard, dut doesn't wake & continues to be in
suspend.

Change-Id: I65befb2fd9085330bca7b8e46ba4e3582fe34af5
Signed-off-by: Poornima Tom <poornima.tom@intel.com>