usb4_gen4 : Support for higher speed cable
As per USB Power Delivery Specification Revision-3.2, Version-1.0,
2023-10, Table 6.40 “UFP VDO” has updated with usb highest speed
for USB4/GEN4. Hence, the UFP VDO is updated with this info.
As this change is missing, when TBT5 cable is used with TR based dock,
it was not detected as a super speed device. With the change in CL,
this issue is fixed.

BUG=b:345881189
BRANCH=None
TEST=Tested on Aviko with TBT5 cable & TR based TBT3 dock.
Device enumerated as TBT.
Port 2: USB=0 DP=0 POLARITY=INVERTED HPD_IRQ=0 HPD_LVL=0 SAFE=0 TBT=1 USB4=0

Change-Id: I7312b4f3846206e5d1d06705af59a6c0ee294389
Signed-off-by: Poornima Tom <poornima.tom@intel.com>