pdc: Apply PD policies following a late sysjump
Following a late sysjump, the AP is already up and running, so we do not
receive the usual hook callbacks to apply the correct policy. This CL
proactively applies the desired policy based on current AP power state
if a late sysjump occurred.

Refactor `enforce_pd_chipset_resume_policy_1()` to set a flag and carry
out the actual steps in `run_snk_policies()` so it can be called before
the PDC Power Mgmt state machine stabilizes.

BUG=b:335700987
TEST=Cannot reproduce problem using procedure described in bug
LOW_COVERAGE_REASON=Add coverage in b/339686045

Change-Id: I5f3b292860c7c2837e20d1a7df95ffeb0ee9ce98