WIP: pdc: Apply PD policies following a late sysjump
Following a late sysjump, the AP is already up and running, so we do not
receive the usual hook callbacks to apply the correct policy. This CL
proactively applies the desired policy based on current AP power state
if a late sysjump occured.

TODO: (1) Test on devices, (2) write unit test.

BUG=b:335700987
TEST=In progress

Change-Id: I5f3b292860c7c2837e20d1a7df95ffeb0ee9ce98