dartmonkey: Change fpinfo host cmd to use fpc_get_hwid
This deduplicates the FPC get HW version spi transaction
functionality and aligns fpc_get_hwid and fp_sensor_get_info
with libbep.

To make this change functionally equivalent to the before state,
we needed to move the global "error" variable update to the one
caller of fpc_get_hwid that expects the global error state to
have been updated.

BUG=b:347111619
TEST=ectool --name=cros_fp fpinfo

Change-Id: I82e5ff2843e68fa9a4ac10fc7ccf889035c1c629