test: pdc: Fix flaky test - test_chipset_resume
This test checks if the PDC's CC operating mode (CCOM) is DRP when the
chipset resume policy is applied. However, this only has an effect if the
port is unattached (see pdc_power_mgmt_set_dual_role()).

Remove the port partner that is connected to the PDC as part of this test
so the PDC stays in an unattached state and carries out the policy.

BUG=b:352404817
TEST=./twister -s pdc/pdc.generic

Change-Id: I5cda7332570adff338e3dd04b184851e07c73007