core/cortex-{m,m0}: specify full path to libclang_rt.builtins-*.a
The directory in which libclang_rt.builtins-*.a resides is no longer
automatically added to the search path in newer Clang versions. As a
result, -lclang_rt.builtins-armv7m fails to find it. With this change,
we instead pass the full path to the library as an input file. This will
cause the library to be used for symbol resolution, without adding its
location to the search path.

BUG=b:346309204
TEST=With LLVM r530567:
     emerge-${BOARD} chromeos-base/chromeos-fpmcu-bloonchipper-unittests

Change-Id: Idd279f763a43e7ac11965367e7be4c94bca84db1