Kconfig.defaults: disable PM_DEVICE_SYSTEM_MANAGED by default
Zephyr power management implementation at the current state has this
weird design inconsistency where the PM subsystem automatically suspends
and resumes all devices with the CPU, if the CPU is using the power
management system. This does not happen if devices implement runtime PM,
or if PM_DEVICE_SYSTEM_MANAGED is disabled.

The problem is that this only makes sense for devices that are somehow
tied to the CPU, but most of the drivers use the device PM callback for
suspend/resume of the device itself, not in support of the CPU power
states. This is the case for the MCU we use in the EC as well, shutting
this option off allows to use device PM APIs for our driver without the
PM subsystem trying to suspend and resume devices when the CPU goes
idle.

BUG=none
TEST=verified on rex, without this device PM gets called on idle

Change-Id: If3290b99a5b66c458de049278da1129d482e5dff