brox: fix power down sequencing
The Intel platform design guide states that non deep-sleep (non-DSx)
systems should drive RSMRST and DSW_PWROK low simultaneously.
However the EC is unable to drive these signals together, so one must
be driven low first.

Swap the order so that RSMRST is always driven to a physical low state
first.  This conforms with Intel's requiresments for DSx enabled systems
and is the safer order when the signals cannot be driven simultaneously.

BUG=b:351912054
TEST=Boot image on brox, run shutdown from AP and from EC

Change-Id: I91f77e509f0ff60dc947398763cfb627604963e4