pdc: Fix public API timeouts
The public API path for PDC commands a loop counter, but calculated the
timeout counter based on the state machine loop delay of 25 ms instead
of the public API delay of 10 ms.

Switch to using k_timepoint_t so that the public API timeout is always
consistent, regardless of the internal delay time changes.

BUG=none
TEST=Run "pdc prs 0" and verify no timeout

Change-Id: I228ec098423a73ac0c96cdc4c02342b8f82f06dc