emul_tps6699x: Handle I2C offset, length correctly
Populate the initial length of the TPS6699x's read response with the
size of the register being read (for now, a theoretical maximum size).
Access the register using an offset that excludes this length byte.
Avoid truncating the first byte of each register read.

1-byte writes contain only the register offset to read in a subsequent
message. Do not treat these as writes to that register.

BUG=b:345292002
TEST=twister -s pdc/pdc.generic.tps6699x
LOW_COVERAGE_REASON=Tests being added in b:345292002

Change-Id: I88abbe28a7177f359d635cd885014194f0eed484