emul_tps6699x: Indicate error on unknown command
If the PPM writes a command that the emulator does not recognize, set
the command register to "!CMD" to indicate an invalid command.
Currently, this applies to almost all commands, which causes tests to
fail, but they fail more obviously than they used to.

BUG=b:345292002
TEST=twister -s pdc/pdc.generic.tps6699x
LOW_COVERAGE_REASON=Tests being added in b:345292002

Change-Id: I18a406d594154b9a7354374770992d736fcd27c3