WIP: usb_dc: fix touchpad data lost and latency issues
This commit adds the touchpad report to a queue if the previous report
hasn't been received by the controller, preventing data loss. The busy
flag is cleared if writing data to the endpoint fails. Additionally, to
resolve the touchpad latency issue, the polling interval is set to 1ms,
ensuring the controller issues an IN token every 1ms to poll for USB
data.

BUG=b:339757474
TEST=on spikyrock, touchpad/keyboard functions work.
TEST=on spikyrock,
     $cat /sys/bus/usb/devices/1-1.1/1-1.1\:1.1/ep_83/interval
     1ms

Change-Id: I4a7f82d36b06a03ac59845e57c7f7f61594ee0f0