usb_dc: fix tp data loss issue and improve tp data bridging latency
This commit adds the touchpad report to a queue if the previous report
hasn't been received by the controller, preventing data loss. The busy
flag is cleared if writing data to the endpoint fails. Additionally, to
improve the touchpad bridging latency, the polling interval is set to
1ms, ensuring the controller issues an IN token every 1ms to poll for
USB data.

BUG=b:339757474
TEST=on spikyrock, touchpad/keyboard functions work.
TEST=on spikyrock,
     $cat /sys/bus/usb/devices/1-1.1/1-1.1\:1.1/ep_83/interval
     1ms
LOW_COVERAGE_REASON=keyboard and touchpad test for Zephyr still not implement

Change-Id: I4a7f82d36b06a03ac59845e57c7f7f61594ee0f0