ec: Enable eigen3 path to be overridden
Introduce a Makefile hook for third_party build variables. The third_party/rules.mk was designed to introduce strictly make rules, which are included too late to be able to influence things like adding build output dirs and source includes for certain step.

The relative path of the eigen3 library was changed after unnesting the
private fingerprint repos. Create a separate variable to enable
overriding the path of the eigen3 library and use it when invoking the
compare_build script.

BUG=b:354273703,b:314208406
TEST=make buildall -j10
TEST=util/compare_build.sh -b all

Cq-Depend: chrome-internal:7516434
Change-Id: Ic031a81f6248907817326ec4c9fb75da7977967a