pdc_power_mgmt: Apply PRS preference via flag
If the sink policy prefers to perform a PRS to become source, express
this by setting a flag to swap to source, which will be applied later to
enqueue an internal command. Do not attempt to perform a blocking public
command directly from the Attached.SNK state.

BUG=b:351255382
TEST=Attach phone; observe DUT initiating PRS upon reboot
LOW_COVERAGE_REASON=Add coverage in b/339686045

Change-Id: Ia094716e86bffc790c6df8485835785e41ffaf5c