pdc: Consistently use enum pdo_source_t
Use enum pdo_source_t to represent the Partner PDO bit in the UCSI
GET_VDOs request and in functions that implement this command in drivers
and emulators. Also use it in emulator set_pdos functions in
anticipation of supporting partner PDOs in emulators.

BUG=b:351255382
TEST=twister -s pdc/pdc.generic
LOW_COVERAGE_REASON=TPS6699x driver coverage to be added in b:345292002

Change-Id: I240816ad516b895157b403f6977ec8617fdf4aa7