zephyr: Enable automatic download of coreboot-sdk using bazel
Coreboot-sdk is now distributable via subtool, and will be removed
from the SDK at /opt/coreboot-sdk.  The easiest way to fetch it is
with Bazel.  This adds logic which fetches coreboot-sdk for the
appropriate architecture when it's not provided by either
COREBOOT_SDK_ROOT or COREBOOT_SDK_ROOT_${ARCH}.

This involved re-factoring how we pass environment variables, as we
now require a number of new environment variables for Bazel to work.
Consequently, I deleted an annoying test that tested the debug
logs (which included the environment variables), as it's simply too
fragile.

BUG=b:338458418
TEST=zmake build -a

Change-Id: I0483554bcf76cea361cc9a1db4f297bc9eeef2b1