util/flash_ec: Fix syntax error due to dump_fmap
Update error handling of dump_fmap invocation to avoid shell syntax
errors when parsing unexpected output. In the error case, dump_fmap
prints some usage text, etc. which should not be relied upon to detect a
valid invocation. It's better to check the command exit status for
failure indication.

Originally, util/flash_ec would fail with a syntax error like

  dump_fmap: unrecognized switch: -e
  util/flash_ec: line 1895: [: 0
  ...

when dump_fmap (a.k.a. futility) is older and does not support the '-e'
flag. This can happen on fast-forward branches where the EC git repo is
head of other branches.

Now, we get:

  ERROR: The output of dump_fmap is not expected; try update_chroot

BUG=none
TEST=checked error message for "util/flash_ec" on nissa FW branch.
	used "util/flash_ec --no_preserve" on nissa FW branch to
	flash the EC.

Change-Id: I823ccb7bfe133a99e620df3cd84bca0fd25b2b72