ucsi: Add logic to intercept SET_SINK_PATH
This CL adds code to intercept the ucsi command SET_SINK_PATH in the
function ucsi_ppm_execute_cmd_sync(). SET_SINK_PATH is used by the
kernel to control charging on a given port and should not be sent by
the PPM directly to the PDC. Instead, SET_SINK_PATH is used to trigger
a call to charge_manager_set_override() which will handle any changes
related to the active charging port.

BUG=b:351728957,b:349685116
TEST=verfied on Brox that the active charging port can be changed
between port 0 and port 1 using
/sys/class/power_supply/<port>/charge_control_limit_max

Change-Id: I17a7f422f13f1fe6ab1494199663f89f55197497