zephyr: llvm: Indicate that Newlib is supported on ARM toolchain only
CL:5297714 enabled TOOLCHAIN_HAS_NEWLIB for all LLVM toolchains. As a
result, tests that require Newlib are not filtered properly.

This was fixed by moving the TOOLCHAIN_HAS_NEWLIB variable from
generic.cmake to target.cmake. It introduces another problem that the
variable is not available while processing Kconfig files, so
CONFIG_NEWLIB_LIBC_SUPPORTED would be never enabled. This was fixed by
overriding the config and enabling it if we are compiling for ARM
architecture.

BUG=b:324574796
TEST=zmake build bloonchipper

Change-Id: Ib2f4cce53f33a7ba5f9419708e8edcbf1792878d