emul_realtek: Add support for SET_PDO sub command
Add support for the SET_PDO sub command. This comamnd is used to set the
LPM's sink or source PDOs.

Also remove the restriction for changing PDO index 0 of the LPM. The
SET_PDO command overwrites all LPM PDOs, starting at offset 0.  The "PDO
Number" field is the count of PDOs sent, not a PDO index.

This also removes the pdo[] array from the GET_PDO request.  The PDO
information is only used in the GET_PDO response.

BUG=none
TEST=./twister -ivc -T zephyr/test/pdc

Change-Id: I587c780ea3276161dbb3c334dbda2bcfabba4eba