pdc_power_mgmt: Add support for HPD Wake
Updates the PDC power manager to track DP status when the connector
status indicates the PDC is in an alternate mode. If HPD LVL switches
from low to high while a device is suspended, the PDC power manager will
wake the AP by signaling a USB mux event.

BUG=b:355452299
TEST=zmake build ${BOARD}, ./twister -T zephyr/test/pdc, manually
confirmed HPD wake was functional.

Change-Id: I9b74a0bc87798a9af4e4f62651fb82ddab849af5