pdc: Implement external external power role swap policy
Currently, the PDC gets configured to allow power role swaps in every
situation. The correct policy is:

 * When powered up (S0): allow external power role swaps in either
   source or sink states.
 * When suspended (S3/S0ix): if we are a source, continue, but only
   allow external swaps into sink.
 * When shut down (S5/G3): same as suspend, but we immediately become a
   sink or disconnect.

The external swap allow bit is changed upon entering the attached-source
or -sink states, and on-demand by setting a policy flag. This is
important for when dualrole mode changes but the PD role stays the same.

Remove an unused function from `pdc_power_mgmt.c`

N.B.: PD roles are from the EC/DUT/PDC's perspective.

BUG=b:359583431
BUG=b:359355196
TEST=Live testing with DUT on top of chromium:5770775
TEST=tast run <IP> firmware.ECPDPowerRole
TEST=./twister -T zephyr/test/pdc

Change-Id: Ia5b1210f64f881ad8a0637d4086de55b456aa6a7