cr50: remove ap ro verification
Disable running AP RO verification on cr50.

This does not change behavior the factory process uses. It leaves
the vendor commands to get/set the hash and AP RO verification result.
It still lets you trigger verification, but the result is always set to
UNSUPPORTED_TRIGGERED. It won't verify the AP RO flash on any devices.
The result is still set back to NOT_RUN when the AP resets.

The factory uses ap_ro_hash.py to set the hash. The script reads the
flash contents from the host, generates the hash, and then sends the
vendor command to save the hash in cr50. This process still works.
This CL doesn't change any of this behavior. The set hash vendor command
is still the same. We plan on keeping it since it's difficult to
update old factory processes to remove a step.

BUG=b:346447095
TEST=manual
get/set the AP RO hash.

Trigger verification. Verify the result changes from NOT_RUN to
UNSUPPORTED_TRIGGERED

Verify the result gets reset to NOT_RUN after the device resets.

Run tast Cr50APROVerification* devboard tests

Change-Id: I63180d6c782b797d5dbb88b5cf91e840c499e2df