pdc: Report chip/driver name in EC_CMD_PD_CHIP_INFO and `pdc info`
Expand the `pdc_info_t` struct to include space for a chip/driver name.
Populate this based on the EC driver's compat string for the RTK and TI
drivers.

Implement V3 of EC_CMD_PD_CHIP_INFO and populate the chip/driver name in
the response struct and display the value in `pdc info`. This simplifies
programmatically determining which PDC series is installed on a device.

BUG=b:356454900
TEST=ectool pdchipinfo 0
TEST=pdc info 0

Change-Id: I50cb37b95b8a958f7d9b95162a09edc01adb1f1f