pdc: Add mechanism to control which PDC ports get FW updates
Our FW implements two-port PDC chips as two separate driver instances
with two separate DT nodes/ports. When Depthcharge queries the EC for PD
chip information, it sees both ports of the PDC individually and will
perform a FW update against each port, but this is unnecessary since
both ports can belong to the same PDC chip. Depthcharge currently has a
hack to block the second update, but this should really be controlled by
information from the EC.

Set up a flag in the EC_CMD_PD_CHIP_INFO response to supress FW updates
for a given port. The EC sets this flag based on a new `no-fw-update`
property inserted into the Devicetree. For dual-port PDCs, this property
should be set on exactly one of the nodes. The flag can also be seen in
the `pdc info` EC console command.

Depthcharge will be updated in a future CL to read this flag and skip
the FW update for that port if set.

BUG=b:324892496
TEST=ectool pdchipinfo 0/1
TEST=pdc info 0/1
TEST=./twister -T zephyr/test/pdc

Change-Id: Ifb7f69700c56f451120234909343305d210c298a