test: Make fetching sensor hwid more generic for fpsensor_hw test
We can't easily use if(IS_ENABLED()) statement blocks with all
code logic, since there is still use of undefined symbols. We could
change this if we allowed including of all driver functions, since
the function names should be scoped as elan_ or fpc_.

Do note that the CONFIG_FP_SENSOR_FPC1025 configs are only enabled
in the RW section, so the assert on the RW is simply for clarity.

BUG=b:347111619
TEST=./test/run_device_tests.py --board=helipilot --with_private=no \
     --flasher=servo_micro --tests fpsensor_hw

Change-Id: I07618832e09b8a9e59076f49524cb9c2323894f5