test/fpsensor_hw: Add buccaneer support
We can't easily use if conditionals on IS_ENABLED for the sensor
selection, since there is still use of undefined symbols. We could
change this if we allowed including of all driver functions, since
the function names should be scoped as elan_ or fpc_.

Do note that the CONFIG_FP_SENSOR_FPC1025 configs are only enabled
in the RW section, so the assert on the RW is simply for clarity.

BUG=b:347111619
TEST=./test/run_device_tests.py --board=helipilot --with_private=no \
     --flasher=servo_micro --tests fpsensor_hw

Change-Id: I878d82445cd18efa70c19f1b395c9ce07ce30acd