common: Fix IS_ENABLED error when present twice in same source
This macro had issues when being used multiple times in one source
file, while being compiled with clang. The complaint is that the
special function attribute, error, was being applied to the same
function, but with different messages. This is because the error
message being applied had the unique CONFIG_ name in it. This might
be fine for GCC, clang was not about it.

You would see the following compilation error:
error: attribute 'error' is already applied with different arguments

Fix Strategy:
* We need to scope the imaginary IS_ENABLED_BAD_ARGS function for each
  use OR we need to remove the CONFIG_ name from the error message.
  I think we want to maintain the CONFIG_ name in the error message.
* We must concatenate something onto "option" in the IS_ENABLED
  definition, to stop later macro expansion from occurring.
* We don't need to add any more unique identifiers, like line number,
  to the function name, since clang only complains about defining
  a *different* error message in the same source file. Since the
  error message only contains the CONFIG_ token name itself, all
  usages of IS_ENABLED for the same CONFIG_ name, in the same source,
  will assign matching error messages. Thus, no warning.
* Using this approach, the imaginary function would be named something
  like IS_ENABLED_BAD_ARGS_CONFIG_FIX_EVERYTHING for
  IS_ENABLED(CONFIG_FIX_EVERYTHING).

BUG=b:361576707, b:346419141
TEST=make host-is_enabled_error
TEST=make run-is_enabled_error
TEST=make BOARD=buccaneer test-timer

Change-Id: I47a669dd426073b898b95ca216f9018e58a3716b