common: Fix IS_ENABLED multiple use with clang
This macro had issues when being used multiple times in one source
file, when being compiled with clang. The complaint is that the
special function attribute, error, was being applied to the same
function, but with different messages, since the message included
the config name.

You would see the following compilation error:
error: attribute 'error' is already applied with different arguments

BUG=b:361576707
TEST=

Change-Id: I47a669dd426073b898b95ca216f9018e58a3716b